refactor: unify offset handling
